Refactored routing in index.py
